Add comma to Harvest tagline
Regardless of formal correctness, as a slogan, this reads much better to me with a comma.  It's punchier and emphasizes the main point, "explore your data."  Even if marketing materials have escaped into the wild without a comma, Harvest is going to be around for a long time, and I think it's worth making the change.