Remove formatting in summary section to prevent it applying to page
When the summary was being truncated using the ellipsis, the trailing formatting end character(** in our case) was lost and the bolding was being applied to the whole page. This is a quick fix and not a long term solution as the code generating the summary should not be truncating formatting characters.

Signed-off-by: Don Naegely <naegelyd@gmail.com>